pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pexams"
version = "0.7.0"
description = "A Python library for generating and correcting exams using Playwright and OpenCV."
authors = [
    { name = "Oscar J. Pellicer-Valero" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-python",
    "numpy",
    "pdf2image",
    "pydantic",
    "Markdown",
    "Faker",
    "Pillow",
    "playwright",
    "matplotlib",
    "pandas",
    "tabulate",
    "torch",
    "torchvision",
    "transformers",
    "timm",
    "pymdown-extensions",
    "python-Levenshtein",
    "pypdf"
]

[project.scripts]
pexams = "pexams.main:main"

[project.urls]
Homepage = "https://github.com/OscarPellicer/pexams"

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["pexams*"]

[tool.setuptools.package-data]
pexams = ["*.css", "assets/sample_test.json", "assets/test_image.jpg", "assets/sample_test.md"]
